Add 'disable_checkpoints' config setting, refactor checkpoint code
- This commit adds a config file option in server settings 'disable_checkpoints'
- If a user sets 'skip_pow_validation' they will skip pow validation on all blocks up to last checkpoint
- If a user sets 'disable_checkpoints' in addition to 'skip_pow_validation', they will fast sync all the way to chaintip
- This also required a relocation BlockchainCheckpoints code, and resolves an omission i made in original addition of this functionality (not sure it would have made a difference, but its resolved here also)
